mcp-server/requirements.txt
fastmcp>=0.1.0
anthropic>=0.25.0
aiofiles>=23.2.1
psutil>=5.9.0
